# docs/swagger.yaml
basePath: /
definitions:
  handler.Error:
    properties:
      message:
        type: string
    type: object
...
  handler.OrderIdResponse:
    properties:
      data:
        $ref: '#/definitions/models.Order'
    type: object
...
  handler.getOrdersResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Order'
        type: array
    type: object
  handler.getWishItemsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Item'
        type: array
    type: object
...
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
...
  models.Item:
    properties:
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      item_created_at:
        type: string
      item_updated_at:
        type: string
      price:
        type: number
      stock:
        type: integer
      title:
        type: string
    required:
    - price
    - title
    type: object
  models.Order:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      item_count:
        description: ordered, delivering, delivered, cancelled
        type: integer
      item_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Документация API для интернет-магазина TechShop
  title: TechShop API
  version: "1.0"
paths:
...
  /api/order:
    get:
      description: Получить все заказы пользователя
      operationId: get-orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getOrdersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Get all orders
      tags:
      - order
  /api/order/{id}:
    get:
      description: Получить конкретный заказ по ID
      operationId: get-order-by-id
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OrderIdResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Get order by ID
      tags:
      - order
  /api/order/{id}/cancel:
    post:
      description: Отменить заказ
      operationId: cancel-order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Cancel order
      tags:
      - order
...
  /api/order/create:
    post:
      description: Создать заказы для всех товаров в корзине
      operationId: create-orders-from-cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Create orders from cart
      tags:
      - order
  /api/order/create/{id}:
    post:
      description: Создать заказ из конкретного товара в корзине
      operationId: create-order-from-cart
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Create order from cart item
      tags:
      - order
  /api/wishlist:
    get:
      description: Получить вишлист пользователя
      operationId: get-
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getWishItemsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Get user's wishlist
      tags:
      - wishlist
  /api/wishlist/{id}:
    delete:
      description: Удалить товар из вишлиста пользователя
      operationId: delete-from-wishlist
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Delete item from wishlist
      tags:
      - wishlist
    post:
      consumes:
      - application/json
      description: Добавить товар в вишлист пользователя
      operationId: add-to-wishlist
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Add item to wishlist
      tags:
      - wishlist
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and your API token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"